port regexp_like function and port related tests
delete useless test

fix chores

change regexp_like

change lock

change format

fix typo

change prost

remove unused

adding usage

adding dependency

adding dependency

dep
